Keep current definition active even after multiple backspaces
Currently, if the user hits "backspace" multiple times, the
active tile will move backwards for each backspace. A backspace
on the first tile of the definition will cause the active definition
to be reset.
After this change, a backspace on the first tile of the definition
will keep it as the active tile. This allows users to repeatedly
hit backspace to delete their current definition solution without
accurately counting the number of backspaces, while keeping the current
definition active.